pyproject.toml
[tool.poetry]
name = "vcc-urn-resolver"
version = "0.1.0"
description = "URN resolver for German administration (OOP reference implementation)"
authors = ["makr-code"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pydantic = "^2.8.0"
SQLAlchemy = "^1.4.0"
alembic = "^2.0.0"
python-dotenv = "^1.0.0"
httpx = "^0.27.2"
psycopg = {version = "^3.2.0", extras = ["binary"]}
pytest = "^8.3.3"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
ruff = "^0.6.9"
black = "^24.10.0"
mypy = "^1.11.2"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = false
warn_return_any = false
warn_unused_configs = true
pretty = true
exclude = [
	"alembic/",
	"tests/",
]

[tool.poetry.scripts]
start = "app.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"